frontend

+++ app/(tabs)/admin.tsx
@@ -44,7 +44,7 @@ import { fetchClubs } from '@/data/clubsRepo';
 function formatDate(iso: string | null): string {
   if (!iso) return '—';
   const d = new Date(iso);
-  return d.toLocaleDateString('en-US', { month: 'short', day: 'numeric', year: 'numeric' });
+  return d.toLocaleDateString('en-US', { month: 'long', day: 'numeric', year: 'numeric' });
 }
 
 // ---------------------------------------------------------------------------
